requirements.txt
instagrapi
pillow
google-generativeai
sqlite3
dotenv
